+++ packages/ui/src/components/indicator/indicator.tsx
@@ -1,0 +1,15 @@
+import type { TBaseTagMap } from "@jay-js/system";
+import { cn } from "../../utils/cn";
+import type { TIndicator } from "./indicator.types";
+
+export function Indicator<T extends TBaseTagMap = "div">(
+	{ tag, className, children, ...props }: TIndicator<T> = { tag: "div" as T },
+): HTMLElementTagNameMap[T] {
+	const Tag = (tag ?? "div") as any;
+
+	return (
+		<Tag {...(props as any)} className={cn("indicator", className)}>
+			{children}
+		</Tag>
+	) as unknown as HTMLElementTagNameMap[T];
+}